reorganized functions from utils
